backend/requirements.txt
# Core Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
sqlalchemy>=2.0.36
psycopg[binary]>=3.1.0
alembic==1.13.0

# Data Validation & Settings
pydantic[email]==2.11.0
pydantic-settings==2.5.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# API & Utilities
python-multipart==0.0.9
python-dotenv==1.0.0

# AI & Agents - OpenAI Agents SDK
openai-agents[litellm]>=0.4.2

# Kafka Event Streaming (Phase V)
aiokafka>=0.10.0

# Push Notifications
pywebpush>=1.14.0

# Date utilities for recurring tasks
python-dateutil>=2.8.2

# Redis for caching (optional)
redis>=5.0.0

